Homework/Chatbot/Lo de Copi/requirements.txt
# LangChain y dependencias principales
langchain==0.2.16
langchain-community==0.2.17
langchain-text-splitters==0.2.4
langchain-openai

# Embeddings y modelos
sentence-transformers
transformers
torch

# Vector Database
faiss-cpu

# Utilidades
numpy==1.26.4
scipy
scikit-learn

# API y configuración
python-dotenv
requests

# Interfaz web
streamlit
streamlit-chat

# PDF processing
pypdf

# Otras utilidades
tqdm
pydantic
